Fix e2e tests with Grafana
