--- .github/scripts/grafana/provisioning/dashboards/symfony.yml
@@ -1,11 +1,0 @@
-apiVersion: 1
-providers:
-  - name: 'Symfony Dashboards'
-    orgId: 1
-    folder: ''
-    type: file
-    disableDeletion: false
-    updateIntervalSeconds: 10
-    allowUiUpdates: true
-    options:
-      path: /etc/grafana/provisioning/dashboards
