--- .github/scripts/grafana/provisioning/datasources/prometheus.yml
@@ -1,8 +1,0 @@
-apiVersion: 1
-datasources:
-  - name: Prometheus
-    type: prometheus
-    access: proxy
-    url: http://prometheus:9090
-    isDefault: true
-    editable: true
